git-config refactoring completed

+++ src/ssp/dl/classifier/naive_text_classifier.py
@@ -1,14 +1,16 @@
 import os
 
 import gin
+import numpy as np
 import pandas as pd
 import tensorflow as tf
 import spacy
 import pickle
-
+import pyarrow as pa
 from ssp.logger.pretty_print import print_info, print_error
 from ssp.utils.misc import check_n_mk_dirs
 
+# https://stackoverflow.com/questions/55854885/unable-to-save-the-tensorflow-model-file-into-hdfs
 
 @gin.configurable
 class NaiveTextClassifier(object):
@@ -23,13 +25,36 @@ class NaiveTextClassifier(object):
                  num_words=8000,
                  seq_len=128,
                  embedding_size=128,
-                 batch_size=64):
+                 batch_size=64,
+                 hdfs_host=None,
+                 hdfs_port=None):
+        """
+
+        :param train_df_or_path:
+        :param test_df_or_path:
+        :param dev_df_or_path:
+        :param model_dir: Local directory path or HDFS path
+        :param wipe_old_data:
+        :param text_column:
+        :param label_column:
+        :param num_words:
+        :param seq_len:
+        :param embedding_size:
+        :param batch_size:
+        """
         self._pre_trained = False
         self._text_column = text_column
         self._label_column = label_column
-        self._train_df = train_df_or_path
-        self._test_df = test_df_or_path
-        self._dev_test = dev_df_or_path
+
+        if isinstance(train_df_or_path, str):
+            self._train_df, self._test_df, \
+            self._dev_test = NaiveTextClassifier.load_parquet_data(train_file_path=train_df_or_path,
+                                                                   test_file_path=test_df_or_path,
+                                                                   dev_file_path=dev_df_or_path)
+        else:
+            self._train_df = train_df_or_path
+            self._test_df = test_df_or_path
+            self._dev_test = dev_df_or_path
 
         self.nlp = spacy.load("en_core_web_sm")
         self._num_words = num_words
@@ -38,28 +63,50 @@ class NaiveTextClassifier(object):
         self._batch_size = batch_size
         self._epochs = 5
         self._threshold = 0.5
-
-        check_n_mk_dirs(model_dir, is_remove=wipe_old_data)
-        self._model_dir = model_dir
-
-        if os.path.exists(self._model_dir + 'tokenizer.pickle'):
-            print_info(f"Loading tokenizer... {self._model_dir + 'tokenizer.pickle'}")
-            with open(self._model_dir + 'tokenizer.pickle', 'rb') as f:
-                self._tokenizer = pickle.load(f)
-        else:
-            self._tokenizer = tf.keras.preprocessing.text.Tokenizer(num_words=self._num_words, oov_token='<UNK>')
-
-        self._model_name = "naive_text_classifier.h5"
-
-        if os.path.exists(self._model_dir + self._model_name):
-            self._pre_trained = True
-            print_info(f"Loading model...{self._model_dir + self._model_name}")
-            self._model = tf.keras.models.load_model(self._model_dir + self._model_name)
-        else:
-            self._model = self.get_model()
-
         self._train_seqs = None
         self._test_seqs = None
+        self._model_name = "naive_text_classifier"
+        self._model_dir = model_dir
+        self._hdfs_host, self._hdfs_port = hdfs_host, hdfs_port
+
+        if hdfs_host and hdfs_port:
+            self._hdfs_fs = pa.hdfs.connect(hdfs_host, hdfs_port)
+            self._is_local_dir = False
+
+            self._model_path = f"{self._hdfs_port}/{self._model_dir}/{self._model_name}".replace("//", "/")
+            self._model_path = f"hdfs://{self._hdfs_host}:" + self._model_path
+
+            if self._hdfs_fs.exists(self._model_dir + 'tokenizer.pickle'):
+                print_info(f"Loading tokenizer... {self._model_dir + 'tokenizer.pickle'}")
+                with self._hdfs_fs.open(self._model_dir + 'tokenizer.pickle', 'rb') as f:
+                    self._tokenizer = pickle.load(f)
+            else:
+                self._tokenizer = tf.keras.preprocessing.text.Tokenizer(num_words=self._num_words, oov_token='<UNK>')
+
+            if self._hdfs_fs.exists(self._model_dir + self._model_name):
+                self._pre_trained = True
+                print_info(f"Loading model...{self._model_path}")
+                self._model = tf.keras.models.load_model(self._model_path)
+            else:
+                self._model = self.get_model()
+        else:
+            check_n_mk_dirs(model_dir, is_remove=wipe_old_data)
+            if os.path.exists(self._model_dir + 'tokenizer.pickle'):
+                print_info(f"Loading tokenizer... {self._model_dir + 'tokenizer.pickle'}")
+                with open(self._model_dir + 'tokenizer.pickle', 'rb') as f:
+                    self._tokenizer = pickle.load(f)
+            else:
+                self._tokenizer = tf.keras.preprocessing.text.Tokenizer(num_words=self._num_words, oov_token='<UNK>')
+
+            if os.path.exists(self._model_dir + self._model_name):
+                self._pre_trained = True
+                print_info(f"Loading model...{self._model_dir + self._model_name}")
+                self._model = tf.keras.models.load_model(self._model_dir + self._model_name)
+            else:
+                self._model = self.get_model()
+
+            self._hdfs_fs = None
+            self._is_local_dir = True
 
     @staticmethod
     def load_parquet_data(train_file_path, test_file_path, dev_file_path):
@@ -67,10 +114,12 @@ class NaiveTextClassifier(object):
                pd.read_parquet(test_file_path, engine="fastparquet"), \
                pd.read_parquet(dev_file_path, engine="fastparquet")
 
-
     def fit_tokenizer(self):
-        # create tokenizer for our data
-        self._tokenizer.fit_on_texts(self._train_df[self._text_column])
+        if self._train_df is not None:
+            # create tokenizer for our data
+            self._tokenizer.fit_on_texts(self._train_df[self._text_column])
+        else:
+            print_info("No data for training!")
 
     def transform(self, text_list):
         # convert text data to numerical indexes
@@ -79,10 +128,14 @@ class NaiveTextClassifier(object):
         seqs = tf.keras.preprocessing.sequence.pad_sequences(seqs, maxlen=self._seq_len, padding="post")
         return seqs
 
-    def preprocess_text(self):
-        self.fit_tokenizer()
-        self._train_seqs = self.transform(text_list=self._train_df[self._text_column])
-        self._test_seqs = self.transform(text_list=self._test_df[self._text_column])
+    def preprocess_train_data(self):
+        if self._train_df is not None:
+            self.fit_tokenizer()
+            self._train_seqs = self.transform(text_list=self._train_df[self._text_column])
+            self._test_seqs = self.transform(text_list=self._test_df[self._text_column])
+        else:
+            print_info("No data for training!")
+
 
     def get_model(self):
         model = tf.keras.Sequential([
@@ -115,12 +168,29 @@ class NaiveTextClassifier(object):
         print(res)
 
     def save_model(self):
-        self._model.save(self._model_dir + self._model_name)
-        # saving
-        with open(self._model_dir + 'tokenizer.pickle', 'wb') as handle:
-            pickle.dump(self._tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)
+        if self._is_local_dir:
+            if os.path.exists(self._model_dir + self._model_name):
+                return
+            self._model.save(self._model_dir + self._model_name)
+            with open(self._model_dir + 'tokenizer.pickle', 'wb') as handle:
+                pickle.dump(self._tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)
+        else:
+            if self._hdfs_fs.exists(self._model_path):
+                return
+            print_info(f"Saving model to {self._model_path}")
+            self._model.save(self._model_path)
+            with self._hdfs_fs.open(self._model_dir + 'tokenizer.pickle', 'wb') as handle:
+                pickle.dump(self._tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)
 
-    def predict(self, text_list):
-        seqs = self.transform(text_list=text_list)
-        preds = self._model.predict(seqs)
-        return preds
+    def predict(self, X):
+        if isinstance(X, list) or isinstance(X, np.ndarray):
+            seqs = self.transform(text_list=X)
+            preds = self._model.predict(seqs)
+            return preds
+        elif isinstance(X, pd.DataFrame):
+            X["prediction"] = self.transform(text_list=X["text"])
+            return X
+        elif isinstance(X, str):
+            res = self.transform([X])
+            return res[0]
+
